+++ src/components/[guild]/JoinModal/components/JoinStep.tsx
@@ -1,20 +1,7 @@
-import {
-  ButtonGroup,
-  ButtonProps,
-  Circle,
-  CircularProgress,
-  HStack,
-  Icon,
-  Spinner,
-  Text,
-  Tooltip,
-  useColorModeValue,
-} from "@chakra-ui/react"
+import { ButtonGroup, ButtonProps, HStack, Text, Tooltip } from "@chakra-ui/react"
 import Button from "components/common/Button"
-import { Check, X } from "phosphor-react"
 import React, { PropsWithChildren } from "react"
-import { getJoinStepIndicatorProps } from "../utils/getStepIndicatorProps"
-import { JoinState } from "../utils/mapAccessJobState"
+import JoinStepIndicator from "./JoinStepIndicator"
 
 type Props = {
   title: string
@@ -27,81 +14,6 @@ type Props = {
   isDone: boolean
   addonButton?: JSX.Element
 } & Omit<ButtonProps, "isDisabled">
-
-const JoinStepIndicator = (
-  props:
-    | { status: "INACTIVE" | "DONE" | "ERROR" | "LOADING" }
-    | { status: "PROGRESS"; progress: number }
-) => {
-  const color = useColorModeValue("black", "white")
-
-  switch (props.status) {
-    case "DONE":
-    case "ERROR":
-    case "INACTIVE": {
-      return (
-        <Circle
-          size="5"
-          border={"1px"}
-          {...(props.status === "DONE"
-            ? {
-                bg: "green.500",
-                borderColor: "green.500",
-              }
-            : props.status === "ERROR"
-            ? { bg: "gray.500", borderColor: "gray.500" }
-            : { bg: "blackAlpha.100", borderColor: "whiteAlpha.100" })}
-        >
-          {(props.status === "DONE" || props.status === "ERROR") && (
-            <Icon
-              as={props.status === "ERROR" ? Check : X}
-              weight="bold"
-              color={"white"}
-              boxSize={"0.8em"}
-              {...(props.status === "DONE"
-                ? { as: Check, boxSize: "0.8em" }
-                : { as: X, boxSize: "0.7em" })}
-            />
-          )}
-        </Circle>
-      )
-    }
-
-    case "LOADING": {
-      return (
-        <Circle border={"1px transparent"}>
-          <Spinner boxSize="5" opacity=".6" />
-        </Circle>
-      )
-    }
-
-    case "PROGRESS": {
-      return (
-        <CircularProgress
-          value={props.progress}
-          size="5"
-          color={color}
-          sx={{
-            "> svg > .chakra-progress__track": {
-              stroke: "var(--chakra-colors-blackAlpha-200)",
-            },
-          }}
-        />
-      )
-    }
-  }
-}
-
-const JoinStatusStep = ({
-  joinState,
-  entity,
-}: {
-  joinState: JoinState
-  entity: "role" | "requirement" | "reward"
-}) => {
-  const props = getJoinStepIndicatorProps(entity, joinState)
-  return <JoinStepIndicator {...props} />
-}
 
 const JoinStep = ({
   title,
@@ -160,5 +72,4 @@ const JoinStep = ({
   )
 }
 
-export { JoinStatusStep, JoinStepIndicator }
 export default JoinStep
