+++ src/components/[guild]/JoinModal/components/ProgressJoinStep/components/JoinStatusStep.tsx
@@ -1,0 +1,16 @@
+import { JoinState } from "components/[guild]/JoinModal/utils/mapAccessJobState"
+import JoinStepIndicator from "../../JoinStepIndicator"
+import { getJoinStepIndicatorProps } from "../utils/getStepIndicatorProps"
+
+const JoinStatusStep = ({
+  joinState,
+  entity,
+}: {
+  joinState: JoinState
+  entity: "role" | "requirement" | "reward"
+}) => {
+  const props = getJoinStepIndicatorProps(entity, joinState)
+  return <JoinStepIndicator {...props} />
+}
+
+export default JoinStatusStep
