+++ src/components/[guild]/JoinModal/components/ProgressJoinStep/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ProgressJoinStep"
