+++ src/components/[guild]/JoinModal/components/SatisfyRequirementsJoinStep.tsx
@@ -1,128 +1,9 @@
-import {
-  Center,
-  Collapse,
-  HStack,
-  Icon,
-  Stack,
-  StackProps,
-  Text,
-  Tooltip,
-} from "@chakra-ui/react"
+import { Center, Collapse, Icon, StackProps, Text, Tooltip } from "@chakra-ui/react"
 import useGuild from "components/[guild]/hooks/useGuild"
 import Button from "components/common/Button"
 import { ArrowRight, LockSimple } from "phosphor-react"
-import { PropsWithChildren, ReactNode } from "react"
 import { JoinState } from "../utils/mapAccessJobState"
-import { JoinStatusStep } from "./JoinStep"
-
-export const JOIN_LOADING_TESTS: Record<
-  Exclude<JoinState["state"], "FINISHED">,
-  [string, string]
-> = {
-  INITIAL: ["Preparing access check", null],
-  PREPARING: [
-    "Preparing access check",
-    "There are a lot of users joining right now, so you have to wait a bit. There're POSITION users before you",
-  ],
-  CHECKING: [null, "Waiting for the next one, there're POSITION users before you"],
-  MANAGING_ROLES: [
-    "Evaluating which roles you have access to",
-    "There are a lot of users joining right now, so you have to wait a bit. There're POSITION users before you",
-  ],
-  MANAGING_REWARDS: [
-    "Evaluating which rewards you will get",
-    "There are a lot of users joining right now, so you have to wait a bit. There're POSITION users before you",
-  ],
-}
-
-const JoinStateCount = ({
-  joinState,
-  entity,
-}: {
-  joinState: JoinState
-  entity: "role" | "reward" | "requirement"
-}) => {
-  if (!joinState) {
-    return null
-  }
-
-  if (entity === "requirement" && joinState.requirements) {
-    if (joinState.state === "CHECKING") {
-      return (
-        <Text>
-          {joinState.requirements.checked}/{joinState.requirements.all} requirements
-          checked
-        </Text>
-      )
-    }
-
-    return (
-      <Text>
-        {joinState.requirements.satisfied}/{joinState.requirements.all} requirements
-        satisfied
-      </Text>
-    )
-  }
-
-  if (entity === "reward" && joinState.rewards) {
-    return (
-      <Text>
-        {joinState.rewards.granted}/{joinState.rewards.all} rewards granted
-      </Text>
-    )
-  }
-
-  if (entity === "role" && joinState.roles) {
-    return (
-      <Text>
-        {joinState.roles.granted}/{joinState.roles.all} roles granted
-      </Text>
-    )
-  }
-
-  return null
-}
-
-const progressTitle = {
-  role: "Get roles",
-  reward: "Get rewards",
-  requirement: "Satisfy the requirements",
-}
-
-const ProgressJoinStep = ({
-  joinState,
-  entity,
-  shouldShowSubtitle,
-  children,
-  RightComponent,
-  ...stackProps
-}: PropsWithChildren<{
-  joinState: JoinState
-  entity: "role" | "reward" | "requirement"
-  shouldShowSubtitle: boolean
-  RightComponent?: ReactNode
-}> &
-  StackProps) => (
-  <HStack py="3" {...stackProps}>
-    <JoinStatusStep entity={entity} joinState={joinState} />
-
-    <Stack w="full" spacing={0} mt="-1.5px !important">
-      <Text fontWeight={"bold"}>{progressTitle[entity]}</Text>
-
-      <JoinStateCount joinState={joinState} entity={entity} />
-
-      {shouldShowSubtitle &&
-        JOIN_LOADING_TESTS[joinState?.state]?.[+!!joinState?.waitingPosition] && (
-          <Text>
-            {JOIN_LOADING_TESTS[joinState?.state][+!!joinState?.waitingPosition]}
-          </Text>
-        )}
-
-      {children}
-    </Stack>
-    {RightComponent}
-  </HStack>
-)
+import ProgressJoinStep from "./ProgressJoinStep"
 
 const SatisfyRequirementsJoinStep = ({
   isLoading,
